+++ json_serializable/test/integration/json_test_example.g_non_nullable.g.dart
@@ -13,6 +13,7 @@ Person _$PersonFromJson(Map<String, dynamic> json) {
     _$enumDecode(_$CategoryEnumMap, json[r'$house']),
     middleName: json['middleName'] as String,
     dateOfBirth: DateTime.parse(json['dateOfBirth'] as String),
+    timesAccessed: json['timesAccessed'] as int ?? 0,
   )
     ..order = Order.fromJson(json['order'] as Map<String, dynamic>)
     ..customOrders = MyList.fromJson((json['customOrders'] as List)
@@ -38,6 +39,7 @@ Map<String, dynamic> _$PersonToJson(Person instance) => <String, dynamic>{
           instance.houseMap.map((k, e) => MapEntry(k, _$CategoryEnumMap[e])),
       'categoryCounts': instance.categoryCounts
           .map((k, e) => MapEntry(_$CategoryEnumMap[k], e)),
+      'readOnly': instance.readOnly,
     };
 
 T _$enumDecode<T>(