+++ json_serializable/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 3.4.1-dev
+## 3.5.0-dev
 
 - Support properties where the getter is defined in a class with a corresponding
   setter in a super type.
+- Add support for the annotations `JsonKey.ignoreDecode` and `JsonKey.ignoreEncode`
 
 ## 3.4.0